Add Mercado Pago payments, credits, and cash admin payments

+++ backend/database/migrations/2026_02_05_140030_create_student_credit_wallets_table.php
@@ -10,7 +10,7 @@ return new class extends Migration
     {
         Schema::create('student_credit_wallets', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('student_id');
+            $table->integer('student_id');
             $table->integer('balance')->default(0);
             $table->timestamps();
